requirements.txt
flask
mysql-connector-python
firebase-admin
werkzeug